requirements-api.txt
# API-specific requirements (no GUI dependencies)

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Core packages
requests==2.32.3
numpy==2.1.3
Pillow==9.4.0

# Computer Vision
opencv-python-headless==4.10.0.84

# PyTorch CPU-only (smaller, faster to download)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
torchvision==0.20.1+cpu

# Face detection/recognition (slowest to install - dlib compiles from source)
dlib==19.24.6
